Adjust the background observer logic and update the file structure
